Enhance Patron ticket functionality with new landing details and integration
- Added a new section in the tickets JSON for the Patron ticket, including labels for row name, action, and notes related to the Scholarship Program and Nepal Relief.
- Updated the PatronCard component to utilize the new usePatronInfo hook for improved state management and data fetching.
- Integrated Patron ticket details into the TicketsPage, ensuring it appears correctly based on availability and pricing structure.
