+++ devcon/content/en/intl/tickets.json
@@ -85,6 +85,14 @@
     "quantity_label": "Number of tickets",
     "checkout": "Continue to checkout",
     "checkout_loading": "Loading…",
+    "landing": {
+      "row_name": "Patron",
+      "row_detail": "Pay what you choose",
+      "row_action": "Choose amount",
+      "row_action_fixed": "Get ticket",
+      "row_note": "A full Devcon ticket whose price funds the Devcon Scholarship Program.",
+      "row_note_relief": "Everything above the minimum goes to Nepal Relief."
+    },
     "page": {
       "back": "Back to Ticket Store",
       "title": "Patron Ticket",
